requirements/base.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
django-rest-framework==0.1.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.2
phonenumbers==9.0.0
pillow==11.0.0
propcache==0.3.0
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
twilio==9.4.6
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
